A good bit of stuff here. Fixed all the bugs discovered by WcaTest.
This involved removing the wcaValue field from our database schema.
Now a missing solve is either null, or undefined, because the solves
array simply isn't long enough for it. This cleaned up a lot of code,
but did require adding more logic to compareSolveTimes. I also had to
change server/exportResults.js to compute the wcaValue, rather than
relying upon the one in the database.