LESS Cleanup.
Renamed bs-constants to ccm-variables so that it  better reflect what the file truly is.
Added a mixins file, so that it's kept separate of the variables.
Moved all of the rules in layout.import.less to be nested - increased readability and less code duplication!
Took advantage of variables as selectors.
Took advantage of clever parent selectors.